+++ sdk/storage/Azure.Storage.DataMovement.Blobs/tests/ProgressHandlerTests.cs
@@ -118,12 +118,12 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
             TransferManagerOptions transferManagerOptions = default,
             TransferOptions transferOptions = default,
             bool trackBytes = true,
-            StorageResourceCreationPreference createMode = StorageResourceCreationPreference.OverwriteIfExists,
+            StorageResourceCreationMode createMode = StorageResourceCreationMode.OverwriteIfExists,
             int waitTime = 30)
         {
             transferManagerOptions ??= new TransferManagerOptions()
             {
-                ErrorHandling = TransferErrorMode.ContinueOnFailure
+                ErrorMode = TransferErrorMode.ContinueOnFailure
             };
 
             TransferManager transferManager = new TransferManager(transferManagerOptions);
@@ -209,9 +209,9 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
         }
 
         [Test]
-        [TestCase(StorageResourceCreationPreference.SkipIfExists)]
-        [TestCase(StorageResourceCreationPreference.FailIfExists)]
-        public async Task ProgressHandler_Conflict(StorageResourceCreationPreference createMode)
+        [TestCase(StorageResourceCreationMode.SkipIfExists)]
+        [TestCase(StorageResourceCreationMode.FailIfExists)]
+        public async Task ProgressHandler_Conflict(StorageResourceCreationMode createMode)
         {
             // Arrange
             using DisposingLocalDirectory source = DisposingLocalDirectory.GetTestDirectory();
@@ -234,8 +234,8 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
                 destinationResource,
                 _expectedBytesTransferred.Take(_expectedBytesTransferred.Length - 2).ToArray(),
                 fileCount: 5,
-                skippedCount: createMode == StorageResourceCreationPreference.SkipIfExists ? 2 : 0,
-                failedCount: createMode == StorageResourceCreationPreference.FailIfExists ? 2 : 0,
+                skippedCount: createMode == StorageResourceCreationMode.SkipIfExists ? 2 : 0,
+                failedCount: createMode == StorageResourceCreationMode.FailIfExists ? 2 : 0,
                 createMode: createMode);
         }
 
@@ -279,7 +279,7 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
 
             TransferManagerOptions transferManagerOptions = new TransferManagerOptions()
             {
-                ErrorHandling = TransferErrorMode.StopOnAnyFailure,
+                ErrorMode = TransferErrorMode.StopOnAnyFailure,
                 MaximumConcurrency = 3
             };
             TransferOptions transferOptions = new TransferOptions()